feat: implement application lifecycle handler
- 어플리케이션 라이프 사이클 핸들러 추가
- 초기 데이터 생성 및 종료 전 데이터 리프레시 기능 구현
